backend/requirements.txt
Django==5.2.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.4.0
django-rq==2.10.2
redis==5.0.3
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.42.0
scikit-learn==1.4.2
vaderSentiment==3.3.2
dj-database-url==2.1.0
psycopg2-binary==2.9.9
